Move utils to proper packages
