+++ app/src/nxt/java/com/pileproject/drive/programming/visual/block/sequence/SetRightMotorSpeedBlock.java
@@ -25,7 +25,7 @@ import com.pileproject.drive.R;
 import com.pileproject.drive.execution.ExecutionCondition;
 import com.pileproject.drive.execution.MachineController;
 import com.pileproject.drive.execution.NxtController;
-import com.pileproject.drive.util.Unit;
+import com.pileproject.drive.util.development.Unit;
 
 /**
  * Set right motor power
