+++ app/src/nxt/java/com/pileproject/drive/programming/visual/block/sequence/StopSecBlock.java
@@ -24,7 +24,7 @@ import android.widget.TextView;
 import com.pileproject.drive.R;
 import com.pileproject.drive.execution.ExecutionCondition;
 import com.pileproject.drive.execution.MachineController;
-import com.pileproject.drive.util.Unit;
+import com.pileproject.drive.util.development.Unit;
 
 import java.util.Locale;
